Uninitialized value $sOptions
In case we have a command line, in the perl script, like:
```
@winchook.
```
we get a message like (line number might be off due to other patches):
```
Use of uninitialized value $sOptions in concatenation (.) or string at /usr/lib/perl5/site_perl/5.18.2/Doxygen/Filter/Perl.pm line 835.
```

In case of a not set `$sOptions` we now see whether a '.' or ',' is present.