Correcting multiple use of doxygen section labels
When having a project with multiple file often messages like:
```
.../RRDs.pm:6: warning: multiple use of section label 'NAME' while adding section, (first occurrence:.../RRDp.pm, line 10)
```
appear. This is due to the fact that that both files have a line:
```
=head1 NAME
```

To overcome this see to it that the create label is unique. This is done based on the fullpath name of the file and a counter.
- Perl.pm
  - transfer the fullpath name to the right (global) place
- Perl.pm
  - adjust file name so we have no "strange" characters in the label
  - adjust the writing of the head1 and head2
  - (for consistency) also translate head3 and head4 to `\subsubpsection` and `\paragraph`