Incorrect end of comment detected
In case, in an item we have something like: `=item 1. Add*/Mod*/Del*/ - high-level` then the `*/` are after translation to cpp seen as end of comment. So they have to be escaped.
(this might happen on other places as well, as soon as found they can be fixed in a similar way).

Note: end of comments in code section will be handled by doxygen in https://github.com/doxygen/doxygen/pull/7026