Doxygen supports `<code>`
There is no need to replace the `<code>` with `\code` as doxygen supports the `<code>`.
In case of e.g.
```
C<Net::FTP::AutoReconnect> is a wrapper module around C<Net::FTP>.
```
this resulted in a number of lines as the `<code>` started a new block and we also got the warning:
```
warning: End of list marker found without any preceding list items
```
due to the `.` as single item starting a line.