Doxygen supports `<code>`
Doxygen supports code so it looked like the view_seq_code was not required but  the double quote has to be escaped.
For the '<pre>'  command the `-` sign at the beginning needs to be escaped otherwise it is seen as the start of a list (it doesn't hurt that the other `-` signs are escaped as well), so we need special handling of verbatim sections.